requirements.txt
# $ conda create --name MMHCAN --file requirements.txt
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.9.9=h06a4308_0
expat=2.7.1=h6a678d5_0
filelock=3.13.1=pypi_0
fsspec=2024.6.1=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.5.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
markupsafe=2.1.5=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h7934f7d_0
networkx=3.3=pypi_0
numpy=2.1.2=pypi_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=9.1.0.70=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-nccl-cu11=2.21.5=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
openssl=3.0.18=hd6dcaed_0
pandas=2.3.3=pypi_0
pillow=11.0.0=pypi_0
pip=25.2=pyhc872135_0
pthread-stubs=0.3=h0ce48e5_1
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
readline=8.3=hc2a1206_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
setuptools=78.1.1=py310h06a4308_0
six=1.17.0=pypi_0
sqlite=3.50.2=hb25bd0a_1
sympy=1.13.3=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.15=h54e0aa7_0
torch=2.7.1+cu118=pypi_0
torchaudio=2.7.1+cu118=pypi_0
torchvision=0.22.1+cu118=pypi_0
triton=3.3.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.2=pypi_0
wheel=0.45.1=py310h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zlib=1.3.1=hb25bd0a_0